requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
pandas==2.0.2
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.30.0
sentry-sdk==1.25.1
six==1.16.0
sqlparse==0.4.4
sympy==1.11.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2